Improvements for CMEM-5922: - Renamed "Input Task attributes" to "Input task attributes" - Pretty print JSON - Fixed typo JSONN -> JSON - Moved new operators to Category "Metadata" instead of "Dataset" - Removed "Dataset parameter" operator - Fixed bug: Now reconfiguring the input task of the new JSON operators works.
