+++ silk-plugins/silk-serialization-json/src/test/scala/org/silkframework/serialization/json/transformer/InputTaskJsonTransformerTest.scala
@@ -13,7 +13,7 @@ import org.silkframework.runtime.serialization.WriteContext
 import org.silkframework.serialization.json.JsonSerialization
 import org.silkframework.serialization.json.JsonSerializers._
 import org.silkframework.workspace.{Project, ProjectTask, TestWorkspaceProviderTestTrait}
-import play.api.libs.json.JsValue
+import play.api.libs.json.{JsValue, Json}
 
 class InputTaskJsonTransformerTest extends AnyFlatSpec with Matchers with TestWorkspaceProviderTestTrait with TestUserContextTrait {
 
@@ -22,7 +22,8 @@ class InputTaskJsonTransformerTest extends AnyFlatSpec with Matchers with TestWo
   it should "retrieve the full JSON of the input dataset" in {
     implicit val writeContext: WriteContext[JsValue] = WriteContext.fromProject(TestData.project)
     val expectedJson = JsonSerialization.toJson[Task[TaskSpec]](TestData.dataset)
-    TestData.retrieve("") shouldBe expectedJson.toString()
+    val actualJson = TestData.retrieve("")
+    actualJson shouldBe Json.prettyPrint(expectedJson)
   }
 
   it should "retrieve a single property of the input dataset" in {
